pyproject.toml
[project]
name = "simple-chat-app"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.71.0",
    "fastapi>=0.120.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
